fix: fix weaviate and opensearch
